Added in more testcases for quicksort.
There is now more sorted and unsorted test cases for quicksort.
There is now also an example of how we can implement the string compare functions that
will be easy to use for end users. Lastly, there is an example test that shows how to reverse
the comparator function and therefore reverse the sort order.