Added in pushing support.
Along with support for pushing things into the heap I had to change how the function generic works.
Instead of using a where at the top of every function, I now use one at the top of the impl. This specifies the type for the entire impl statement and means I don't have to
keep typing it out. Also I made it so unimplemented tests will fail to make it more obvious what else
I have to do.